Add files via upload
This is the code I ran to control my arm. The code initializes the servos from the highest port down setting them to positions in an array. The array allows you to simply enter the number of servo you would like to control whenever, then you begin to move it. You need to have potentiometers connected to successfully control the servos with this code.